Refactor API-server database unit tests.
Distinguish between database initialization and database migration unit tests.

Prepares for #3130.
